refactor: split helper functions into different .py
